refactor data loading + projects
