+++ src/data/projects.json
@@ -1,0 +1,27 @@
+{
+    "Project1": {
+        "name": "Project1",
+        "description": "Project1 description",
+        "url": "https://github.com/shalankwa/portfolio-site"
+    },
+    "Project2": {
+        "name": "Project2",
+        "description": "Project2 description",
+        "url": "https://github.com/shalankwa/portfolio-site"
+    },
+    "Project3": {
+        "name": "Project3",
+        "description": "Project3 description",
+        "url": "https://github.com/shalankwa/portfolio-site"
+    },
+    "Project4": {
+        "name": "Project4",
+        "description": "Project4 description",
+        "url": "https://github.com/shalankwa/portfolio-site"
+    },
+    "Project5": {
+        "name": "Project5",
+        "description": "Project5 description",
+        "url": "https://github.com/shalankwa/portfolio-site"
+    }
+}
