--- src/data/skills.js
@@ -1,3 +1,0 @@
-import skillsJson from './skills.json'
-
-export const skills = new Map(Object.entries(skillsJson))
